Fixed a few bugs in cleaning pipeline.

+++ pipeline/finalize_labelme_cleaning.sh
@@ -16,6 +16,7 @@ Usage:
      --in_version IN_VERSION
      --subversion SUBVERSION
      --out_version OUT_VERSION
+     --up_to_now UP_TO_NOW
 
 Example:
   $PROGNAME
@@ -34,14 +35,18 @@ Options:
   --out_version
       (optional) If specified, will be promoted to this version. 
                  Otherwise, will use in_version + 1.
+  --up_to_now
+      (optional) 0 or 1. If 1, will finalize only data from all campaigns.
+      If 0, will finalize both campaign_id and all campaigns. Default is 0. 
 EO
 }
 
 ARGUMENT_LIST=(
     "campaign_id"
     "in_version"
-    "out_version"
     "subversion"
+    "out_version"
+    "up_to_now"
 )
 
 opts=$(getopt \
@@ -78,6 +83,10 @@ while [[ $# -gt 0 ]]; do
             out_version=$2
             shift 2
             ;;
+        --up_to_now)
+            up_to_now=$2
+            shift 2
+            ;;
         --) # No more arguments
             shift
             break
@@ -102,15 +111,12 @@ if [ -z "$out_version" ]; then
   out_version=$((${in_version} + 1))
   echo "Argument 'out_version' is not provided. Will use ${out_version}."
 fi
-if (( "$out_version" <= $in_version )); then
-  echo "Argument 'out_version'=$out_version must be > 'in_version'=$in_version."
-  exit 1
-fi
 
 echo "campaign_id:            ${campaign_id}"
 echo "in_version:             ${in_version}"
-echo "out_version:            ${out_version}"
 echo "subversion:             ${subversion}"
+echo "out_version:            ${out_version}"
+echo "up_to_now:              ${up_to_now}"
 
 # The end of the parsing code.
 ################################################################################
@@ -123,32 +129,36 @@ source ${dir_of_this_file}/../constants.sh
 ## Promote subversion to the next version for THIS campaign.
 ## Example: v7.2 will be promoted to v8.
 
-# The path to promote.
-in_6Kx4K_path=$(ls -1 $(get_6Kx4K_db_path ${campaign_id} ${in_version}.${subversion}) | tail -n 1)
-if [ -z "$in_6Kx4K_path" ]; then
-  echo "Failed to set in_6Kx4K_path. Exit."
-  exit 1
+# If ONLY this campaign was cleaned.
+if [ ${up_to_now} -eq 0 ]; then
+
+  # The path to promote.
+  in_6Kx4K_path=$(ls -1 $(get_6Kx4K_db_path ${campaign_id} ${in_version}.${subversion}) | tail -n 1)
+  if [ -z "$in_6Kx4K_path" ]; then
+    echo "Failed to set in_6Kx4K_path. Exit."
+    exit 1
+  fi
+
+  # The new paths.
+  out_6Kx4K_path=$(get_6Kx4K_db_path ${campaign_id} ${out_version})
+  echo "Linking ${in_6Kx4K_path} to ${out_6Kx4K_path}"
+  rm -f ${out_6Kx4K_path}
+  ln -s ${in_6Kx4K_path} ${out_6Kx4K_path}
+
+  # The path to promote.
+  in_1800x1200_path=$(ls -1 $(get_1800x1200_db_path ${campaign_id} ${in_version}.${subversion}) | tail -n 1)
+  if [ -z "$in_1800x1200_path" ]; then
+    echo "Failed to set in_1800x1200_path. Exit."
+    exit 1
+  fi
+
+  # The new paths.
+  out_1800x1200_path=$(get_1800x1200_db_path ${campaign_id} ${out_version})
+  echo "Linking ${in_1800x1200_path} to ${out_1800x1200_path}"
+  rm -f ${out_1800x1200_path}
+  ln -s ${in_1800x1200_path} ${out_1800x1200_path}
+
 fi
-
-# The new paths.
-out_6Kx4K_path=$(get_6Kx4K_db_path ${campaign_id} ${out_version})
-echo "Linking ${in_6Kx4K_path} to ${out_6Kx4K_path}"
-rm -f ${out_6Kx4K_path}
-ln -s ${in_6Kx4K_path} ${out_6Kx4K_path}
-
-# The path to promote.
-in_1800x1200_path=$(ls -1 $(get_1800x1200_db_path ${campaign_id} ${in_version}.${subversion}) | tail -n 1)
-if [ -z "$in_1800x1200_path" ]; then
-  echo "Failed to set in_1800x1200_path. Exit."
-  exit 1
-fi
-
-# The new paths.
-out_1800x1200_path=$(get_1800x1200_db_path ${campaign_id} ${out_version})
-echo "Linking ${in_1800x1200_path} to ${out_1800x1200_path}"
-rm -f ${out_1800x1200_path}
-ln -s ${in_1800x1200_path} ${out_1800x1200_path}
-
 
 ## Promote subversion to the next version for ALL campaigns.
 ## Example: v7.2 will be promoted to v8.
